added more few_shot_examples
