+++ requirements.txt
@@ -1,3 +1,5 @@
 openai>=1.3.0
 pandas>=2.0.0
 python-dotenv>=1.0.0
+pytest>=7.0.0
+unittest-mock>=1.3.0
